LayoutTests/editing/selection/ios/clear-selection-after-tapping-on-element-with-click-handler-expected.txt

This test verifies that the DOM selection is dismissed when tapping on an element that listens to click events. To manually test, select 'WebKit' and tap on the red square. The selection should be dismissed, and the output area should indicate that no text is selected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().toString() is ""
PASS getSelection().type is "Caret"
PASS successfullyParsed is true

TEST COMPLETE
WebKit
The selected text is: ""
